+++ src/layer/reduction.cpp
@@ -1041,7 +1041,7 @@ static int reduction(const Mat& a, Mat& b, float v0, bool reduce_w, bool reduce_
     if (ret != 0)
         return -100;
 
-    if (post_process || fabs(coeff - 1.f) > FLT_EPSILON)
+    if (post_process || fabsf(coeff - 1.f) > FLT_EPSILON)
     {
         ret = reduction_post_process<Op3>(b, coeff, opt);
         if (ret != 0)
@@ -1065,7 +1065,7 @@ struct post_process_sqrt
 {
     T operator()(const T& x) const
     {
-        return static_cast<T>(sqrt(x));
+        return static_cast<T>(sqrtf(x));
     }
 };
 
@@ -1074,7 +1074,7 @@ struct post_process_log
 {
     T operator()(const T& x) const
     {
-        return static_cast<T>(log(x));
+        return static_cast<T>(logf(x));
     }
 };
 
@@ -1101,7 +1101,7 @@ struct reduction_op_asum
 {
     T operator()(const T& x, const T& y) const
     {
-        return static_cast<T>(x + fabs(y));
+        return static_cast<T>(x + fabsf(y));
     }
 };
 
@@ -1119,7 +1119,7 @@ struct reduction_op_sumsexp
 {
     T operator()(const T& x, const T& y) const
     {
-        return static_cast<T>(x + exp(y));
+        return static_cast<T>(x + expf(y));
     }
 };
